Made changes to shopping basket project

+++ tddpractice/src/main/java/com/qa/basket/item/Item.java
@@ -26,6 +26,6 @@ public class Item {
      *
      */
     public double getTotalCost() {
-        return this.cost * (double) this.quantity;
+        return this.cost * this.quantity;
     }
 }
